Parameterize the tests over implementations.
This lets hypothesis work separately on each implementation, giving it
more time to find problems, and also lets the number of failing tests
(1/12) better match the number of problems in the package, as compared
to the previous 1/4.  The found failure is stull just
`FFTToeplitz(complex)`, at least after fixing the other tests.